ci: add GitHub Actions workflow; make tree fmt- and clippy-clean
Add .github/workflows/ci.yml running fmt --check, clippy -D warnings, and the
full test suite on push/PR, plus a second job exercising the `db` feature.
Stable toolchain only (no MSRV pin). README gets a CI badge.

Bringing clippy to -D warnings surfaced latent lints across the workspace;
fix them all (let-chains for collapsible ifs, needless borrows on CST
node/token refs, redundant must_use on must_use-returning fns, sort_by_key,
unwrap_or_default, print_literal, redundant cast) and apply rustfmt repo-wide.
No behavior change.

+++ crates/mold_lsp/src/semantic_tokens.rs
@@ -61,7 +61,7 @@ pub fn semantic_tokens(root: &SyntaxNode, index: &LineIndex) -> SemanticTokens {
         };
         let range = token.text_range();
         let start = index.position(u32::from(range.start()));
-        let length = (u32::from(range.end()) - u32::from(range.start())) as u32;
+        let length = u32::from(range.end()) - u32::from(range.start());
         // Multi-line tokens (block comments) are rare; encode against the start
         // line only, which is acceptable for highlighting.
         let line = start.line;
